refactor(coverage): extract TestCoverageCoordinator from TestCoverageTools
`RunTestCoverageCore` was a 216 LOC / CC 20 method bundling gating, dotnet
invocation, partition logic, cobertura parsing, and envelope shaping into one
gate-lambda body. Split the pure data-in/data-out helpers out into a new
`TestCoverageCoordinator` service under `RoslynMcp.Roslyn.Services`:

- `PartitionTestProjectsByCoverlet` (+ `TestProjectPartition` record)
- `ParseAndAggregateCoberturaXml` (+ `WeightedAverage`)
- `BuildCoverletMissingResult`, `BuildNoCoverageFileResult`,
  `BuildTimeoutResult`, `BuildUnexpectedErrorResult` failure factories

`RunTestCoverageCore` is now a thin orchestrator (~70 LOC) that delegates to
the coordinator and to a new `RunCoveragePassAsync` helper for the
classic-vs-partial dotnet-test branches. `SerializeWithDeprecation` stays in
the tool class because `JsonDefaults` is internal to `RoslynMcp.Host.Stdio`.

Closes backlog row `test-coverage-tools-runtestcoveragecore-split`.
